+++ src/nuklear_math.c
@@ -200,7 +200,7 @@ nk_log10(double n)
 NK_LIB float
 nk_roundf(float x)
 {
-    return (x >= 0.0) ? nk_ifloorf(x + 0.5) : nk_iceilf(x - 0.5);
+    return (x >= 0.0f) ? (float)nk_ifloorf(x + 0.5f) : (float)nk_iceilf(x - 0.5f);
 }
 NK_API struct nk_rect
 nk_get_null_rect(void)
